+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Hike ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added `obsidian` as a command that the command line understands.
+  ([#41](https://github.com/davep/hike/pull/41))
+
 ## v0.5.0
 
 **Released: 2025-02-25**